Remove CreateDelegation from utxo set
