+++ chainstate/tx-verifier/src/transaction_verifier/timelock_check.rs
@@ -17,7 +17,7 @@ use chainstate_types::{block_index_ancestor_getter, GenBlockIndex};
 use common::{
     chain::{
         block::timestamp::BlockTimestamp, signature::Transactable, timelock::OutputTimeLock,
-        ChainConfig, GenBlock,
+        ChainConfig, GenBlock, TxOutput,
     },
     primitives::{BlockDistance, BlockHeight, Id},
 };
@@ -28,6 +28,20 @@ use super::{
     error::ConnectTransactionError, storage::TransactionVerifierStorageRef,
     TransactionSourceForConnect,
 };
+
+fn get_timelock(output: &TxOutput) -> Option<&OutputTimeLock> {
+    match output {
+        TxOutput::Transfer(_, _)
+        | TxOutput::Burn(_)
+        | TxOutput::StakePool(_)
+        | TxOutput::ProduceBlockFromStake(_, _)
+        | TxOutput::CreateDelegationId(_, _)
+        | TxOutput::DelegateStaking(_, _, _) => None,
+        TxOutput::LockThenTransfer(_, _, tl)
+        | TxOutput::DecommissionPool(_, _, _, tl)
+        | TxOutput::SpendShareFromDelegation(_, _, _, tl) => Some(tl),
+    }
+}
 
 fn check_timelock(
     source_block_index: &GenBlockIndex,
@@ -96,7 +110,7 @@ where
             .map_err(|_| utxo::Error::ViewRead)?
             .ok_or(ConnectTransactionError::MissingOutputOrSpent)?;
 
-        if let Some(timelock) = utxo.output().timelock() {
+        if let Some(timelock) = get_timelock(utxo.output()) {
             let height = match utxo.source() {
                 utxo::UtxoSource::Blockchain(h) => *h,
                 utxo::UtxoSource::Mempool => match tx_source {
