+++ common/src/chain/transaction/output/mod.rs
@@ -65,33 +65,3 @@ pub enum TxOutput {
     #[codec(index = 8)]
     SpendShareFromDelegation(Amount, Destination, DelegationId, OutputTimeLock),
 }
-
-impl TxOutput {
-    pub fn is_burn(&self) -> bool {
-        match self {
-            TxOutput::Transfer(_, _)
-            | TxOutput::LockThenTransfer(_, _, _)
-            | TxOutput::CreateStakePool(_)
-            | TxOutput::DecommissionPool(_, _, _, _)
-            | TxOutput::ProduceBlockFromStake(_, _)
-            | TxOutput::CreateDelegationId(_, _)
-            | TxOutput::DelegateStaking(_, _, _)
-            | TxOutput::SpendShareFromDelegation(_, _, _, _) => false,
-            TxOutput::Burn(_) => true,
-        }
-    }
-
-    pub fn timelock(&self) -> Option<&OutputTimeLock> {
-        match self {
-            TxOutput::Transfer(_, _)
-            | TxOutput::Burn(_)
-            | TxOutput::CreateStakePool(_)
-            | TxOutput::ProduceBlockFromStake(_, _)
-            | TxOutput::CreateDelegationId(_, _)
-            | TxOutput::DelegateStaking(_, _, _) => None,
-            TxOutput::LockThenTransfer(_, _, tl)
-            | TxOutput::DecommissionPool(_, _, _, tl)
-            | TxOutput::SpendShareFromDelegation(_, _, _, tl) => Some(tl),
-        }
-    }
-}
